Move sched api package
There was a duplicate api package in the namespace, just moved it around.
