+++ src/main/java/com/animedetour/android/database/event/FetchedEventMetrics.java
@@ -8,7 +8,7 @@
  */
 package com.animedetour.android.database.event;
 
-import com.animedetour.api.sched.api.model.Event;
+import com.animedetour.api.sched.model.Event;
 import com.j256.ormlite.dao.Dao;
 import com.j256.ormlite.stmt.PreparedQuery;
 import com.j256.ormlite.stmt.QueryBuilder;
