+++ src/main/java/com/animedetour/android/database/favorite/FavoriteRepository.java
@@ -9,7 +9,7 @@
 package com.animedetour.android.database.favorite;
 
 import com.animedetour.android.schedule.favorite.Favorite;
-import com.animedetour.api.sched.api.model.Event;
+import com.animedetour.api.sched.model.Event;
 import com.inkapplications.groundcontrol.SingleYieldWorker;
 import com.j256.ormlite.dao.Dao;
 import com.j256.ormlite.stmt.DeleteBuilder;
