+++ src/main/java/com/animedetour/android/schedule/DayFragment.java
@@ -10,12 +10,10 @@ package com.animedetour.android.schedule;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.ListView;
 import butterknife.Bind;
 import com.animedetour.android.R;
@@ -24,7 +22,7 @@ import com.animedetour.android.framework.BaseFragment;
 import com.animedetour.android.main.SpinnerOptionContainer;
 import com.animedetour.android.main.SubNavigationSelectionChange;
 import com.animedetour.android.schedule.serach.EventSearchActivity;
-import com.animedetour.api.sched.api.model.Event;
+import com.animedetour.api.sched.model.Event;
 import com.inkapplications.android.widget.listview.ItemAdapter;
 import com.inkapplications.groundcontrol.SubscriptionManager;
 import com.squareup.otto.Subscribe;
