+++ src/main/java/com/animedetour/android/schedule/favorite/Favorite.java
@@ -8,7 +8,7 @@
  */
 package com.animedetour.android.schedule.favorite;
 
-import com.animedetour.api.sched.api.model.Event;
+import com.animedetour.api.sched.model.Event;
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
