+++ src/main/java/com/animedetour/api/sched/ScheduleEndpoint.java
@@ -6,9 +6,9 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-package com.animedetour.api.sched.api;
+package com.animedetour.api.sched;
 
-import com.animedetour.api.sched.api.model.Event;
+import com.animedetour.api.sched.model.Event;
 import retrofit.Callback;
 import retrofit.http.GET;
 import retrofit.http.Query;
