+++ src/test/java/com/animedetour/api/sched/api/deserialization/PanelDateDeserializerTest.java
@@ -6,7 +6,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-package com.animedetour.api.sched.api.deserialization;
+package com.animedetour.api.sched.deserialization;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
