+++ src/test/java/com/animedetour/api/sched/api/model/EventTest.java
@@ -6,7 +6,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-package com.animedetour.api.sched.api.model;
+package com.animedetour.api.sched.model;
 
 import junit.framework.TestCase;
 import org.junit.Test;
